validations var

+++ app/Http/Controllers/Admin/PostController.php
@@ -17,11 +17,14 @@ class PostController extends Controller
             'string',
             'max:100',
         ],
-        'title'     => 'required|string|max:100',
-        'image'     => 'url|max:100',
+        'title'         => 'required|string|max:100',
+        'category_id'   => 'required|integer|exists:categories,id',
+        'tags'          => 'array',
+        'tags.*'        => 'integer|exists:tags,id',
+        'image'         => 'url|max:100',
         'uploaded_img'  => 'nullable|image|max:1024',
-        'content'   => 'string',
-        'excerpt'   => 'string',
+        'content'       => 'nullable|string',
+        'excerpt'       => 'nullable|string',
     ];
 
 
@@ -63,37 +66,31 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title'         => 'required|string|max:100',
-            'slug'          => 'required|string|max:100|unique:posts',
-            'category_id'   => 'required|integer|exists:categories,id',
-            'image'         => 'url|max:100',
-            'uploaded_img'  => 'nullable|image|max:1024',
-            'content'       => 'nullable|string',
-            'excerpt'       => 'nullable|string',
-        ]);
+       // validation
+       $this->validations['slug'][] = 'unique:posts';
+       $request->validate($this->validations);
 
+       $data = $request->all();
 
-        $data = $request->all();
+       $img_path = isset($data['uploaded_img']) ? Storage::put('uploads', $data['uploaded_img']) : null;
 
-        $img_path = isset($data['uploaded_img']) ? Storage::put('uploads', $data['uploaded_img']) : null;
+       // salvare i dati nel db
+       $post = new Post;
+       $post->slug             = $data['slug'];
+       $post->title            = $data['title'];
+       $post->category_id      = $data['category_id'];
+       $post->image            = $data['image'];
+       $post->uploaded_img     = $img_path;
+       $post->content          = $data['content'];
+       $post->excerpt          = $data['excerpt'];
+       $post->save();
 
+       // associamo il post appena creato ai tag
+       $post->tags()->attach($data['tags']);
 
-        // salvare i dati nel db
-        $post = new Post;
-        $post->slug             = $data['slug'];
-        $post->title            = $data['title'];
-        $post->category_id      = $data['category_id'];
-        $post->image            = $data['image'];
-        $post->uploaded_img     = $img_path;
-        $post->content          = $data['content'];
-        $post->excerpt          = $data['excerpt'];
-        $post->save();
+       // ridirezionare (e non ritornare una view)
+       return redirect()->route('admin.posts.show', ['post' => $post]);
 
-        // associamo il post appena creato ai tag
-        $post->tags()->attach($data['tags']);
-
-        return redirect()->route('admin.posts.show', ['post' => $post]);
     }
 
     /**
@@ -134,33 +131,29 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $request->validate([
-            'slug'      => [
-                'required',
-                'string',
-                'max:100',
-                Rule::unique('posts')->ignore($post),
-            ],
-            'title'     => 'required|string|max:100',
-            'image'     => 'url|max:100',
-            'uploaded_img'  => 'image|max:1024',
-            'content'   => 'nullable|string',
-            'excerpt'   => 'nullable|string',
-        ]);
+        // validation
+        $this->validations['slug'][] = Rule::unique('posts')->ignore($post);
+        $request->validate($this->validations);
 
         $data = $request->all();
 
-        $img_path = Storage::put('uploads', $data['uploaded_img']);
-        Storage::delete($post->uploaded_img);
+        if (isset($data['uploaded_img'])) {
+            $img_path = Storage::put('uploads', $data['uploaded_img']);
+            Storage::delete($post->uploaded_img);
+        } else {
+            $img_path = $post->uploaded_img;
+        }
 
         // salvare i dati nel db
-        $post->slug     = $data['slug'];
-        $post->title    = $data['title'];
-        $post->image    = $data['image'];
+        $post->slug         = $data['slug'];
+        $post->title        = $data['title'];
+        $post->image        = $data['image'];
         $post->uploaded_img  = $img_path;
         $post->content  = $data['content'];
         $post->excerpt  = $data['excerpt'];
         $post->update();
+
+        $post->tags()->sync($data['tags']);
 
         // ridirezionare (e non ritornare una view)
         return redirect()->route('admin.posts.show', ['post' => $post]);
